# swagger.yaml
openapi: 3.0.3
info:
  title: Whispers Chat Backend
  version: 2.0.0
  description: |
    Chat backend secured with OIDC Authorization Code Flow with PKCE and bearer-token authorization.

    HTTP base URL example: `http://localhost:8080`
    WebSocket URL example: `ws://localhost:8080/ws/user`

    Protected HTTP endpoints require `Authorization: Bearer <access_token>`.
    Refresh uses an `HttpOnly` cookie issued by the backend on login and rotation.

    Browser WebSocket clients cannot reliably send `Authorization` headers, so the
    WebSocket handshake may authenticate either with:
    - `Authorization: Bearer <access_token>`
    - `Sec-WebSocket-Protocol: whispers.bearer.<access_token>`
servers:
  - url: http://localhost:8080
    description: Local development server
security:
  - bearerAuth: []
paths:
...
  /private-chats:
    post:
      summary: Create a PRIVATE browser-tied chat with another user
      description: |
        Creates a new `PRIVATE` chat for the authenticated user's browser `keyId` and the counterpart browser key.
        If a `PRIVATE` chat already exists for the same user pair and the same browser-key pair, that chat is returned.
      parameters:
        - name: targetUserId
          in: query
          required: true
          schema:
            type: string
          example: bob@example.com
        - name: keyId
          in: query
          required: true
          schema:
            type: string
          example: alice-browser-key
      responses:
        '200':
          description: PRIVATE chat created or returned if the same browser-key-bound chat already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatSummary'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /private-chats/{chatId}:
    get:
      summary: Get PRIVATE chat metadata and browser-bound key material references
      parameters:
        - name: chatId
          in: path
          required: true
          schema:
            type: integer
            format: int64
          example: 456
        - name: keyId
          in: query
          required: true
          schema:
            type: string
          example: alice-browser-key
      responses:
        '200':
          description: PRIVATE chat metadata for the browser key bound to that chat
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PrivateChatView'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /private-chats/{chatId}/messages:
    get:
      summary: Get encrypted message history for a PRIVATE chat bound to a specific browser key
      parameters:
        - name: chatId
          in: path
          required: true
          schema:
            type: integer
            format: int64
          example: 456
        - name: keyId
          in: query
          required: true
          schema:
            type: string
          example: alice-browser-key
      responses:
        '200':
          description: Stored encrypted messages for the PRIVATE chat
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PrivateMessageView'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque access token
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Missing, invalid, revoked, or expired bearer token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: Conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    ChatSummary:
      type: object
      required:
        - chatId
        - username
        - type
      properties:
        chatId:
          type: integer
          format: int64
          example: 123
        username:
          type: string
          example: bob@example.com
        type:
          $ref: '#/components/schemas/ChatType'
    ChatType:
      type: string
      enum: [DIRECT, PRIVATE]
      example: DIRECT
...
    PrivateChatKeyStatus:
      type: string
      enum: [ACTIVE, REVOKED]
      example: ACTIVE
    PrivateChatKeyView:
      type: object
      required:
        - keyId
        - publicKey
        - algorithm
        - format
        - status
        - createdAt
        - updatedAt
      properties:
        keyId:
          type: string
          example: alice-browser-key
        publicKey:
          type: string
          example: MIIBIjANBgkqhkiG9w0BAQEFAAOCAQ8A...
        algorithm:
          type: string
          example: RSA-OAEP
        format:
          type: string
          example: spki
        status:
          $ref: '#/components/schemas/PrivateChatKeyStatus'
        createdAt:
          type: string
          example: '2026-04-20T12:00:00Z'
        updatedAt:
          type: string
          example: '2026-04-20T12:01:00Z'
    PrivateChatView:
      type: object
      required:
        - chatId
        - username
        - type
        - currentUserKey
        - counterpartKey
      properties:
        chatId:
          type: integer
          format: int64
          example: 456
        username:
          type: string
          example: bob@example.com
        type:
          $ref: '#/components/schemas/ChatType'
        currentUserKey:
          $ref: '#/components/schemas/PrivateChatKeyView'
        counterpartKey:
          $ref: '#/components/schemas/PrivateChatKeyView'
    EncryptedPrivateMessagePayload:
      type: object
      required:
        - protocolVersion
        - encryptionAlgorithm
        - keyWrapAlgorithm
        - ciphertext
        - nonce
        - senderKeyId
        - senderMessageKeyEnvelope
        - recipientKeyId
        - recipientMessageKeyEnvelope
      properties:
        protocolVersion:
          type: string
          example: v1
        encryptionAlgorithm:
          type: string
          example: AES-GCM
        keyWrapAlgorithm:
          type: string
          example: RSA-OAEP
        ciphertext:
          type: string
          example: base64-ciphertext
        nonce:
          type: string
          example: base64-nonce
        senderKeyId:
          type: string
          example: alice-browser-key
        senderMessageKeyEnvelope:
          type: string
          description: Wrapped symmetric message key for the sender's own browser key
          example: base64-sender-envelope
        recipientKeyId:
          type: string
          example: bob-browser-key
        recipientMessageKeyEnvelope:
          type: string
          description: Wrapped symmetric message key for the recipient browser key
          example: base64-recipient-envelope
...
    PrivateMessageView:
      type: object
      required:
        - chatId
        - senderUsername
        - chatType
        - encryptedMessage
        - timestamp
      properties:
        chatId:
          type: integer
          format: int64
          example: 456
        senderUsername:
          type: string
          example: alice@example.com
        chatType:
          $ref: '#/components/schemas/ChatType'
        encryptedMessage:
          $ref: '#/components/schemas/EncryptedPrivateMessagePayload'
        timestamp:
          type: string
          example: '2026-04-20T12:00:00Z'
...
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          example: Invalid or expired bearer token